Restructures the Tanach component to account for Torah text, and moves the JSON's to the public directory in anticipation of the future
