+++ src/components/Perakim/TorahPerek.jsx
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import TeacherCard from './TeacherCard.jsx';
 import ReaderCard from './ReaderCard.jsx';
+import Tanach from '../Tanach/Tanach.jsx';
 import gematriya from '../../../lib/gematriya.js';
 
 class TorahPerek extends Component {
@@ -20,87 +21,6 @@ class TorahPerek extends Component {
       partNumber: i,
       partTitle: title,
     });
-  }
-
-  showSefer() {
-    const hebArr = this.props.act.hebrew_text || [];
-    const engArr = this.props.act.english_text || [];
-    const start_chapter = this.props.act.start_chapter || null;
-    const start_verse = this.props.act.start_verse || null;
-    const end_chapter = this.props.act.end_chapter || null;
-    const end_verse = this.props.act.end_verse || null;
-
-    let gen = Date.now();
-
-    const heb = [];
-    const eng = [];
-    hebArr.forEach((perek, i) => {
-      const perek_num = start_chapter != null ? start_chapter + i : i + 1;
-      const passuk_num = i === 0 && start_verse != null ? start_verse : 1;
-
-      heb.push(<b key={gen++}>פרק {gematriya(perek_num)}</b>)
-      perek.forEach((passuk, j) => heb.push(<p key={gen++}><b>{gematriya(passuk_num + j)}. </b>{passuk}</p>));
-    });
-    engArr.forEach((perek, i) => {
-      const perek_num = start_chapter != null ? start_chapter + i : i + 1;
-      const passuk_num = i === 0 && start_verse != null ? start_verse : 1;
-
-      eng.push(<b key={gen++}>Perek {perek_num}</b>)
-      perek.forEach((passuk, j) => eng.push(<p key={gen++}><b>{passuk_num + j}. </b>{passuk}</p>));
-    });
-    let par = [];
-    for (let i = 0; i < hebArr.length && i < engArr.length; i++) {
-      const perek_num = start_chapter != null ? start_chapter + i : i + 1;
-      const passuk_num = i === 0 && start_verse != null ? start_verse : 1;
-
-      let currHeb = hebArr[i];
-      let currEng = engArr[i];
-      par.push(
-        <div key={gen++}>
-          <div className='row valign-wrapper hide-on-small-only'>
-            <div className='col l6 m12 s12 left-align valign'><b>Perek {perek_num}</b></div>
-            <div className='col l6 m12 s12 rtl right-align valign'><b>פרק {gematriya(perek_num)}</b></div>
-          </div>
-          <div className='hide-on-med-and-up'>
-            <div className='row center-align rtl'><b>פרק {gematriya(perek_num)}</b></div>
-            <div className='row center-align'><b>Perek {perek_num}</b></div>
-          </div>
-        </div>
-      );
-      for(let j = 0; j < currHeb.length && i < currEng.length; j++) {
-        par.push(
-          <div key={gen++}>
-            <div className='row valign-wrapper hide-on-small-only'>
-              <div className='col l6 m12 s12 left-align valign'><p><b>{passuk_num + j}. </b>{currEng[j]}</p></div>
-              <div className='col l6 m12 s12 rtl right-align valign'><p><b>{gematriya(passuk_num + j)}. </b>{currHeb[j]}</p></div>
-            </div>
-            <div className='hide-on-med-and-up'>
-              <div className='row center-align rtl'><p><b>{gematriya(passuk_num + j)}. </b>{currHeb[j]}</p></div>
-              <div className='row center-align'><p><b>{passuk_num + j}. </b>{currEng[j]}</p></div>
-            </div>
-          </div>
-        );
-      }
-    }
-    if (this.state.show === 'heb') {
-      return (
-        <div className='right-align rtl' id='hebText'>
-          {heb}
-        </div>
-      );
-    } else if (this.state.show === 'par') {
-      return (
-        <div id='parText'>
-          {par}
-        </div>
-      );
-    } else if (this.state.show === 'eng') {
-      return (
-        <div className='left-align' id='engText'>
-          {eng}
-        </div>
-      );
-    }
   }
 
   render() {
@@ -234,7 +154,12 @@ class TorahPerek extends Component {
           <div className='row'>
             <div className='card'>
               <div className='card-content'>
-                {this.showSefer()}
+                <Tanach
+                  sefer={sefer}
+                  perek={perek}
+                  show={this.state.show}
+                  part={act.part_name}
+                />
               </div>
             </div>
           </div>
